requirements.txt
yt-dlp
openai-whisper
torch
transformers
sentencepiece
Flask
